requirements.txt
llama-cpp-python==0.2.83
transformers==4.42.4
torch==2.3.1
streamlit==1.37.0